Move server functions to separate submodule
